% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-data.R
\docType{data}
\name{activity}
\alias{activity}
\title{Data: Activity}
\format{An object of class \code{numeric} of length 1000.}
\usage{
activity
}
\description{
Artificial data. Approximately normally distributed covariate measuring activity of
1000 individuals.
}
\keyword{datasets}
