% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-data.R
\docType{data}
\name{agecategory}
\alias{agecategory}
\title{Age category}
\format{An object of class \code{factor} of length 1000.}
\usage{
agecategory
}
\description{
Artificial data. Five-level factor giving age category of 1000 individuals
}
\keyword{datasets}
