% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox.R
\docType{package}
\name{fwerapprox}
\alias{fwerapprox}
\alias{fwerapprox-package}
\title{The fwerapprox package}
\description{
(1) Computes score test statistics for testing whether a large number of
coefficients (typically corresponding to genetic markers) in a GLM are zero in
presence of a smaller number of covariates (typically environmental covariat) not
included in the test, and provides estimates of correlations beween the test
statistics, and (2) computes Glaz–Johnson-type intersection approximations of the
asymptotic multivariate normal distribution of the test statistics. See Halle et al.,
"Efficient and powerful familywise error control in genome-wide association studies
using generalised linear models".
}
\details{
Use \code{library(help = "fwerapprox")} to get a list of functions and data sets.
}
