% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-functions.R
\name{gamma2}
\alias{gamma2}
\title{Second-order product-type approximation of multivariate normal probablity}
\usage{
gamma2(alphaloc, corr, tol = 1e-07)
}
\arguments{
\item{alphaloc}{Local significance level (se above).}

\item{corr}{A vector of first-order correlations (i.e., between \eqn{T_j} and \code{T_{j+1}}
for \eqn{j = 1}, \ldots, \eqn{m - 1}).}

\item{tol}{Maximal order of which correlations of score test statistics
are estimated.}

\item{tol}{If \eqn{||\rho| - 1| \leq} \code{tol} for a correlation \eqn{\rho}, then the
corresponding factor in the approximation is set to one.}
}
\value{
\eqn{P(|T_1| < c, \ldots, |T_m| <
c)} is approximated for \eqn{(T_1, \ldots, T_m)} multivariate standard normal
with first-order correlations given by \code{corr}, where \eqn{c} is
\code{qnorm(1 - alphaloc/2)}. If \eqn{(T_1,\ldots T_m)} is a test statistic vector
for \eqn{m} hypotheses, a local significance level of \code{alphaloc}, i.e. rejection of
a null hypothesis if the \eqn{p}-value is less than \code{alphaloc}, will control
familywise error rate at the \code{1 - gamma2(alphaloc, corr)} level.
}
\description{
Computes a second-order Glaz--Johnson approximation to a multivariate
standard normal probability with a given correlation matrix. Gives a familywise error rate level
bound in multiple testing for a given local (per-hypothesis) significance level.
}
