% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-functions.R
\name{gamma_k}
\alias{gamma_k}
\title{Order \eqn{k} product-type approximation of multivariate normal probablity}
\usage{
gamma_k(k, alphaloc, corr, miwasteps = 4096, genz = FALSE)
}
\arguments{
\item{k}{Order of the approximation.}

\item{alphaloc}{Local significance level (se above).}

\item{corr}{Either a list of correlations
in which the
first component is a vector of first-order correlations, the second is a vector
of second-order correlations,
up to a vector of \code{k} \eqn{- 1}-order correlations, or a correlation matrix.
Typically the \code{corrs} or \code{corrmatrix} component given by \code{\link{scorestatcorr}}.}

\item{miwasteps}{The \code{steps} parameter used for \code{Miwa} (\code{\link[mvtnorm]{algorithms}})
in \code{\link[mvtnorm]{pmvnorm}}.}

\item{genz}{If \code{TRUE}, the \code{GentzBretz} algorithm (\code{\link[mvtnorm]{algorithms}}) is used
for \code{\link[mvtnorm]{pmvnorm}}, otherwise the \code{Miwa} algorithm is used.}
}
\value{
\eqn{P(|T_1| < c, \ldots, |T_m| <
  c)} is approximated for \eqn{(T_1, \ldots, T_m)} multivariate standard normal
  with covariances (at least up to order \code{k-1}) given by \code{corr}, where \eqn{c} is
  \code{qnorm(1 - alphaloc/2)}. If \eqn{(T_1,\ldots T_m)} is a test statistic vector
  for \eqn{m} hypotheses, a local significance level of \code{alphaloc}, i.e. rejection of
  a null hypothesis if the \eqn{p}-value is less than \code{alphaloc}, will control
  familywise error rate at the \code{1 - gamma_k(alphaloc, corr)} level.
}
\description{
Computes an order \eqn{k} Glaz--Johnson approximation to a multivariate
standard normal probability with a given correlation matrix. Gives a familywise error rate level
bound in multiple testing for a given local (per-hypothesis) significance level.
}
\details{
The function is quite slow. For second-order approximation, it is better to use \code{\link{gamma2}}.
}
\examples{
# Normal model with three environmental covariates:
result <- scorestatcorr(y_normal ~ sex + activity + agecategory, xg[,1:200], 2)
al <- uniroot(function(a) gamma2(a, result$corrs[[1]]) - .95, c(1e-5, 5e-4), tol = 1e-14)$root
al3 <- uniroot(function(a) gamma_k(3, a, result$corrs) - .95, c(1e-5, 5e-4), tol = 1e-14)$root
0.05/2000 # Bonferroni
1 - 0.95^(1/2000) # Sidak
al # order 2 FWER approximation
al3 # order 3 FWER approximation
}
