% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-functions.R
\name{scorestatcorr}
\alias{scorestatcorr}
\title{Score test statistics in generalized linear models, with estimated
correlations}
\usage{
scorestatcorr(formula, xg, maxorder, family = gaussian, both = FALSE)
}
\arguments{
\item{formula}{Model \code{\link[stats]{formula}} for the null model, i.e.
including environmental covariates.}

\item{xg}{Matrix of genetic markers (one column for each marker).}

\item{maxorder}{Maximal order of which correlations of score test statistics
are estimated.}

\item{family}{Family of GLM passed to \code{\link[stats]{glm}}.}

\item{both}{If \code{TRUE}, correlations are given both as a list and as a
matrix (see 'Value').}
}
\value{
A list containing the following components: \item{statistic}{A vector
  of score test statistics} \item{corrs}{A list of vectors of correlations of
  score test statistics. The first component is a vector of first-order
  correlations, i.e. between neighbouring markers, the second is a vector of
  second-order correlations, i.e. between markers of distance 2, etc. The
  number of components is \code{maxorder}. The list is generated if
  \code{maxorder} is greater than 0 and less than the number of columns of
  \code{xg} or if \code{both} is \code{TRUE}.} \item{corrmatrix}{Estimated
  correlation matrix of the score test statistics. The matrix is generated if
  \code{maxorder} is equal to the number of columns of \code{xg} or if
  \code{both} is \code{TRUE}. In the latter case, any correlations between
  markers of distance greater than \code{maxorder} will be set to 0.
  \code{scorestatcorr} will attempt to load the \code{Matrix} package and
  return \code{corrmatrix} as a sparse banded matrix.}.
}
\description{
Computes score test statistics for testing whether each of a large number of
coefficients (typically corresponding to genetic markers) in a GLM with canonical
link is zero
in presence of a smaller number of covariates (typically environmental
covariates) not included in the test, and provides estimates of correlations
between the test statistics.
}
\examples{
# Normal model with three environmental covariates:
result <- scorestatcorr(y_normal ~ sex + activity + agecategory, xg, 2)
result$statistic[1:10]
result$corrs[[1]][1:10]
pvals <- 2*pnorm(-abs(result$statistic)) # p-values for two-sided test
pvals[1:10]

# Logistic model without environmental covariates:
resultl <- scorestatcorr(y_logistic ~ 1, xg, 2, family = binomial)
resultl$statistic[1:10]
resultl$corrs[[1]][1:10]
pvalsl <- 2*pnorm(-abs(resultl$statistic))
pvalsl[1:10]
}
