% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-data.R
\docType{data}
\name{sex}
\alias{sex}
\title{Sex}
\format{An object of class \code{factor} of length 1000.}
\usage{
sex
}
\description{
Artificial data. Two-level factor giving sex of 1000 individuals
}
\keyword{datasets}
