% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-data.R
\docType{data}
\name{xg}
\alias{xg}
\title{Genetic markers}
\format{An object of class \code{matrix} with 1000 rows and 2000 columns.}
\usage{
xg
}
\description{
Artificial data. Matrix containing markers (0, 1, 2, corresponding to genotypes aa, aA, AA)
at 2000 markers (columns) for 1000 individuals (rows).
}
\details{
The frequencies of 0, 1 and 2 are approximately 0.81, 0.18 and 0.01, respectively, at each
marker. Sample correlation beween neighbouring markers is approximately 0.6.
}
\keyword{datasets}
