% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-data.R
\docType{data}
\name{y_logistic}
\alias{y_logistic}
\title{Data: Bernoulli response}
\format{An object of class \code{integer} of length 1000.}
\usage{
y_logistic
}
\description{
Artificial data. Bernoulli response for 1000 individual, with effect of the first genetic
marker (first column of \code{xg}).
}
\keyword{datasets}
