% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-data.R
\docType{data}
\name{y_normal}
\alias{y_normal}
\title{Data: Normal response}
\format{An object of class \code{numeric} of length 1000.}
\usage{
y_normal
}
\description{
Artificial data. Normally distributed response for 1000 individuals, with some effect of
covariates \code{activity}, \code{agecategory} and \code{sex}, and of the first genetic
marker (first column of \code{xg}).
}
\keyword{datasets}
