% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerapprox-data.R
\docType{data}
\name{yl}
\alias{yl}
\title{Bernoulli response}
\format{An object of class \code{integer} of length 1000.}
\usage{
yl
}
\description{
Artificial data. Bernoulli response for 1000 individual, with effect of the first genetic
marker (first column of \code{xg}).
}
\keyword{datasets}
